[fix/#14] Controller에서 Repository Service를 호출하지 않도록 수정
