+++ src/main/java/com/clokey/server/domain/history/application/HistoryService.java
@@ -1,0 +1,6 @@
+package com.clokey.server.domain.history.application;
+
+public interface HistoryService {
+
+    void changeLike(Long memberId, Long historyId, boolean isLiked);
+}
